EFUN: strptime() and strftime()
