+++ docs/apply/index.md
@@ -192,6 +192,6 @@ title: APPLY
 
 <div class="alert alert-info my-4" role="alert">
     <img src="https://cdnjs.cloudflare.com/ajax/libs/octicons/8.5.0/svg/info.svg">
-    This page is auto generated on 2021-02-19 21:35:03 EST for v2019.20201121-33-gc72b462b.</a>
+    This page is auto generated on 2021-02-23 15:50:53 PST for v2019.20201121-53-g74a2c94b.</a>
 </div>
