+++ docs/efun/system/strftime.md
@@ -1,0 +1,21 @@
+---
+layout: default
+title: system / strftime
+---
+
+### NAME
+
+    strftime() - format time as string
+
+### SYNOPSIS
+
+    string strftime( string fmt, int time );
+
+### DESCRIPTION
+
+    Format given time with a format string, see https://en.cppreference.com/w/cpp/chrono/c/strftime
+
+### SEE ALSO
+
+    strptime(3)
+
