+++ docs/efun/system/strptime.md
@@ -1,0 +1,21 @@
+---
+layout: default
+title: system / strptime
+---
+
+### NAME
+
+    strptime() - parse string as local time
+
+### SYNOPSIS
+
+    int strptime( string fmt, string str );
+
+### DESCRIPTION
+
+    Parse given time 'str' with a format string, see https://en.cppreference.com/w/cpp/io/manip/get_time
+
+### SEE ALSO
+
+    strftime(3)
+
