+++ docs/lpc/index.md
@@ -84,6 +84,6 @@ title: lpc
 
 <div class="alert alert-info my-4" role="alert">
     <img src="https://cdnjs.cloudflare.com/ajax/libs/octicons/8.5.0/svg/info.svg">
-    This page is auto generated on 2020-12-13 18:10:15 PST for v2019.20201121-14-g833c706e.</a>
+    This page is auto generated on 2021-02-23 15:50:53 PST for v2019.20201121-53-g74a2c94b.</a>
 </div>
