+++ testsuite/single/tests/efuns/strftime.c
@@ -1,0 +1,4 @@
+void do_tests() {
+  ASSERT(stringp(strftime("%m/%d/%Y %H:%M:%S %Z", 1602454371))); // Sun, 11 Oct 2020 22:12:51 GMT
+  ASSERT_NE(0, strftime("%mmm", 1602454371));
+}
