+++ testsuite/single/tests/efuns/strptime.c
@@ -1,0 +1,4 @@
+void do_tests() {
+  ASSERT(strptime("%m/%d/%Y %H:%M:%S", "10/11/2020 16:12:51") > 0);
+  ASSERT(catch(strptime("%1234", "whatever")));
+}
